+++ src/weight_system.h
@@ -72,6 +72,6 @@ const WeightSystem intersect(const WeightSystem &q1, const WeightSystem &q2,
 
 bool has_ip(const WeightSystem &ws);
 
-constexpr unsigned weight_system_storage_size = dim * sizeof(uint32_t);
+constexpr size_t weight_system_storage_size = dim * sizeof(uint32_t);
 
 #endif
